ZFS-TESTS: Modify Test Suite to include dstat
As part of the effort to provide an analysis capability for the performance
regression test suite dstat is added as a first step to provide the information
for performance analsys. In future phases this will replace iostat, mpstat, and
other tools for gathering information. The patch also changes fio to output to
a json file in addition to its redirected output and renames the output files
to make them easier to parse the embedded components in the path.

Signed-Off-By: David Quigley <david.quigley@intel.com>